requirements.txt
django-crispy-forms
crispy-bootstrap5
gunicorn
whitenoise
Pillow